Carry paintshop over to repair tab
